PR test results are not updated by testgrid/updater
The updater does not know how to handle the `.../directory/*/NUM.txt`
entries and omits those results in `listBuilds()`. Have the updater
read `x-goog-meta-link` and resolve those entries as builds.

At the same time, correct gcsupload to set `x-goog-meta-link` as
bootstrap.py does. This ensures jobs that switch to using gcsupload
to set metadata can be scraped for the testgrid.