kubernetes-csi: generate jobs for all repos, blacklist branches
This makes it possible to run the presubmit job for all current repos,
without automatically starting it. This makes it possible to
experiment with the job in more scenarios.

Once we let it run automatically, blacklisting of branches will
prevent running the job for branches that don't support Prow.

Testing of the currently published sidecars is added also for the
latest Kubernetes 1.13.x, to detect issues with the next Kubernetes
maintenance release. As with master, Kubernetes gets built from source
for that.