testgrid/release: replace kubeadm-gce* jobs with kubeadm-kind(er)
The kubeadm-gce* jobs are no longer supported due to the
underlying deployer - kubernetes-anywhere.
SIG Cluster Lifecycle spent a lot of effort this cycle
implementing a solution for testing kubeadm upgrade and skew
scenarios.

This resulted in a tool called kinder, which is based on the
kind backend but with a few more commands to support
upgrade and skew.

This PR does the following in release-x-all/informing dashboards:
- replace kubeadm-gce-x with kubeadm-kind-x
- replace kubeadm-gce-x-on-y with kubeadm-kinder-y-on-x
- replace kubeadm-gce-upgrade-x-y with kubeadm-kinder-upgrade-x-y