Updating config.proto.
Also adding a warning note about updating config/state.proto before
updating the other version of those files.

I think since this option is not used, changing its type
and name is safe, but I'm not 100% sure on that.

(The replaced field will need to be added first in the other version of
config.proto, and then added back here.)