Continue reprocessing results until they are no longer running.
Currently if jobs take more than 4 hours we will never find their final result.
Avoid this by finding the oldest downloaded column that is still running. Drop it and everything more recent than it.
This will allow up to 24 hours for a job to complete (after which time the overall cell will automatically fail with a timeout message.