Improve sleep message.
The sleeping wait=5m is generally erroneous.
The time spent doing the update is subtracted from the wait time.
This is a substantial amount of time, often more than the wait period.
This means the next update happens much quicker.

Add the until field to make it more obvious when sleeping should stop.