Implement ignore_old_results.
This options results in a group of column only using the most recent result for a given row.
By default there will be a row, row [1], row [2], etc. When this option is on we will instead
just take the most recent cell and put it into row, dropping the other cells.