Limit config.proto fields to those we support.
There are a large number of fields that non-Google testgrid users do not have access to.
This is a combination of fields which are irrelevant to the prow-based updater,
as well as potentially relevant fields but which have not implemented internally.

Simplify the external-facing proto by limiting the test group config to the set of fields that
are relevant externally