Refactor writes to return new object metadata.
When we lock a dashboard by cloud copying it over itself it is useful to track the new generation.
This allows us to conditionally write the new state with the assumption is hasn't changed.

Also, update LeastRecentlyUpdated to set a ceiling on the number of concurrent Stats.
Attempting to stat all paths concurrently is not optimally efficient and can run into
open file limits on some systems.