Refactor updater to continually send groups to be updated.
Currently updater will wait until all groups have been updated once before updating them a second time.
Instead groups are put into a priority queue based on how recently they were updated.
As soon as an item is sent for an update, its moved to the back of the queue.

This helps mitigate the situation where 99% of the groups have been updated but the last one is taking a long time.
Instead of waiting for this one to finish, the other groupConcurrency workers will being updating the other groups
while this one is finishing.