Use buffered channels instead of concurrency.
Currently start many go threads to read artifacts and builds in parallel. This causes a lot of
memory pressure. So instead switch to channels with a 1 item buffer. This allows each go routine
to continue working on the second item as soon as it finishes the first one. Things will only block
if the second item is completed before the first item is read.

Assuming everything is processing quickly this should result in similar performance but with better
performance when things slow down (memory consumed for 1-2 items instead of N).

Also: reject junit files more than 100MB, failing these builds with a note about the error.