Check preconditions when reading.
The official storage client does this incorrectly: it uses the XML api for reads but sends ifGenerationMatch
parameters (for the json api) rather than the required x-goog-if-generation-match headers. So instead check
this manually when the response comes back.

Also fix some related issues (cannot use the testgroup's conditions to read objects, similar thing for reading
the config).