Synthetic columns are not readColumns errors.
Initially the updater aborted the test group when it encountered a problematic build.
Next it would just ignore the column.
Today it emits a synthetic column that gives information about what parts of the build
are problematic.

This obviates the need to have readColumns() return an error indicating whether or not it found
any error-free columns.

Today so long as 1 column is free of errors we return a nil error (even if 9 others have problems).
This changes it so that we'll also return nil error it that last column also contains an error.
Since we always return a nil error, remove this from the function signature.