Allocate half the time to compile and upload results.
We want to allocate half the time to reading new columns and the second half to compiling them.

This fixes this by calculating the time to read new columns at the beginning of the function instead
of after many long operations (downloading old grid, reading first new column).

Currently the effect is more allocate half the time remaining after reading the first column to reading additional columns
and the second half to uploading. This generally works but not in some edge cases where it takes a long
time to get the first column.