TestGrid updater - on dynamic email list, solve casting []string issue
When reading list of strings from json go recognize it as a list of interfaces, which cause casting issues.
Added to metadata function that get list of strings if the type is list of strings or list of interfaces that represent list of strings, and return if the key is exist.