Stop touching the grid before download, tracking generation.
Download and Upload now returns the generation. We can use this in InflateDropAppend to determine what version we downloaded
and ensure when we try and replace it that no one else has changed it.

This is faster and also simplifies Update since it no longer needs to try and lock the object and/or track the generation.

We can just determine the generation when we download the object and ensure that our upload replaces this version.