Prevent concurrently updating the same dashboard.
If 3 tabs get new results at the same time, we only want to update the summary
once, not three times concurrently. So only allow one workers to actively update
a dashboard. However, if a tab is updated while the dashboard is getting updated
then it is likely already stale, so we want to wait for this dashboard to complete
and immediately queue up a new run once it finishes.