Fixing a panic when recording column headers in alerts.
A recent update to Summarizer records column header values
(specifically, the 'extra' or 'custom' column header values) when
recording metadata about alerted rows. We missed a case where, if there
are fewer values recorded than the current number of custom column
headers configured for the tab, we hit an out of index issue. (This is
a legitimate case that can happen when new custom column headers are
added).

Quick-fixed this by checking the length of column header values before
recording them, and an additional unit test to verify this case. (We
made this change internally, and I've verified it works already).