Add ability to query a single job from Prow in ResultStore.
Our ResultStore implementation at the moment is limited to Prow results
from ResultStore (e.g. has "label:prow"). This is fine for our current
assumptions, but if results for multiple Prow jobs are uploaded to the
same project (a common use case), then all the results get processed in
TestGrid on the same tab, which we want to avoid for users and
development.

Instead, allow a very simple query (only `target:"<job name>"`), so it's
possible to specify a particular job for a tab instead of the results
for everything uploaded from Prow.