Report missing podinfo.json as PASS_WITH_SKIPS.
It's not expected that podinfo.json is always uploaded from Prow, or
that it's an error if it doesn't update. Instead of a neutral status
(which affects the tab status in TestGrid), mark this as passing with
skips so there's still an informational message, but it doesn't flag the
tab as failing or flaky.