Use SearchConfiguredTargets when searching targets.
If searching for a target (as opposed to labels or other identifiers),
we want to search for those results using a difference method
(SearchConfiguredTargets, rather than SearchInvocations). Add the search
method and use it if the query contains a target atom (and update the
Prow label accordingly for this search).

Moved the query code into a separate package so it can be used in config
validation without a circular dependency.