Specify auto-project detection for PubSub.
In pubsub v1.32, a breaking change made it so specifying a blank project
results in an error, rather than automatically detecting the project as
it did before. Instead, specify the sentinel value for detecting the
project automatically.

See
https://github.com/googleapis/google-cloud-go/releases/tag/pubsub%2Fv1.32.0.

Note that an update a while back upgraded the pubsub module version from
v1.22.2 to v1.33.0.